+++ src/main/java/diet/dietmanager/DietManagerParser.java
@@ -1,0 +1,4 @@
+package diet.dietmanager;
+
+public class DietManagerParser {
+}
