+++ src/main/java/diet/dietmanager/DietManagerUI.java
@@ -1,0 +1,4 @@
+package diet.dietmanager;
+
+public class DietManagerUI {
+}
