+++ src/main/java/diet/dietmanager/command/NewDietSession.java
@@ -1,6 +1,6 @@
 package diet.dietmanager.command;
 
-public class NewDietSession implements Command{
+public class NewDietSession implements Command {
     @Override
     public void execute(String[] args) {
