+++ src/main/java/diet/dietsession/DietSession.java
@@ -1,0 +1,10 @@
+package diet.dietsession;
+
+public class DietSession {
+    public boolean endWorkoutSession;
+    public static void workoutSessionStart(){
+        while(true) {
+
+        }
+    }
+}
